Changed float to decimal where money

+++ Assets/Scripts/Queue.cs
@@ -22,9 +22,9 @@ public class Queue : MonoBehaviour
     public int _queueLength = 1;
     private int _lengthOwned = 1;
 
-    private double _nextLengthUpgradePrice;
-    private double _nextWaitingTimeUpgradePrice;
-    private double _nextSpawnrateUpgradePrice;
+    private decimal _nextLengthUpgradePrice;
+    private decimal _nextWaitingTimeUpgradePrice;
+    private decimal _nextSpawnrateUpgradePrice;
 
     private int _waitingTimeUpgradesOwned = 0;
     private int _spawnrateUpgradesOwned = 0;
@@ -66,9 +66,9 @@ public class Queue : MonoBehaviour
 
     private void Update()
     {
-        _nextLengthUpgradePrice = System.Math.Round(Constants.QUEUE_LENGTH_UPGRADE_BASE_COST * Mathf.Pow(Constants.MULTIPLIER, _lengthOwned), 2);
-        _nextWaitingTimeUpgradePrice = System.Math.Round(Constants.QUEUE_TIME_UPGRADE_BASE_COST * Mathf.Pow(Constants.MULTIPLIER, _waitingTimeUpgradesOwned), 2);
-        _nextSpawnrateUpgradePrice = System.Math.Round(Constants.QUEUE_SPAWN_UPGRADE_BASE_COST * Mathf.Pow(Constants.MULTIPLIER, _spawnrateUpgradesOwned), 2);
+        _nextLengthUpgradePrice = System.Math.Round((decimal)Constants.QUEUE_LENGTH_UPGRADE_BASE_COST * (decimal)Mathf.Pow(Constants.MULTIPLIER, _lengthOwned), 2);
+        _nextWaitingTimeUpgradePrice = System.Math.Round((decimal)Constants.QUEUE_TIME_UPGRADE_BASE_COST * (decimal)Mathf.Pow(Constants.MULTIPLIER, _waitingTimeUpgradesOwned), 2);
+        _nextSpawnrateUpgradePrice = System.Math.Round((decimal)Constants.QUEUE_SPAWN_UPGRADE_BASE_COST * (decimal)Mathf.Pow(Constants.MULTIPLIER, _spawnrateUpgradesOwned), 2);
         if(IsHandlingPerson() == true && HP_running == false)
         {
             StartCoroutine(HandlePerson());
@@ -152,7 +152,7 @@ public class Queue : MonoBehaviour
     {
         if (GameManager.coins >= _nextLengthUpgradePrice)
         {
-            GameManager.coins -= (float)_nextLengthUpgradePrice;
+            GameManager.coins -= _nextLengthUpgradePrice;
             _lengthOwned++;
             _queueLength++;
 
@@ -174,7 +174,7 @@ public class Queue : MonoBehaviour
     {
         if (GameManager.coins >= _nextWaitingTimeUpgradePrice)
         {
-            GameManager.coins -= (float)_nextWaitingTimeUpgradePrice;
+            GameManager.coins -= _nextWaitingTimeUpgradePrice;
             _waitingTimeUpgradesOwned++;
             _waitingDuration--;
         }
@@ -184,7 +184,7 @@ public class Queue : MonoBehaviour
     {
         if (GameManager.coins >= _nextSpawnrateUpgradePrice)
         {
-            GameManager.coins -= (float)_nextSpawnrateUpgradePrice;
+            GameManager.coins -= _nextSpawnrateUpgradePrice;
             _spawnrateUpgradesOwned++;
             _spawnTime--;
         }
